split kinematic_constraint.h into sub files.
